Refactor: changing folder structure and translating from spanish to english
